Allow any kind of background
When making afba28d0d7a5e36a312ce37795988bc5b5ce6758, I forgot that background-color and background are compatible if one of them are just a fallback, and as a result, ended up removing the ability to set colors as the background (which was the point of the commit to begin with). This commit reverts that mistake.